requirements.txt
argparse>=1.2.1
backports.ssl-match-hostname>=3.4.0.2
docker-py>=0.6.0
execnet>=1.2.0
ipython>=2.3.0
pep8>=1.5.7
py>=1.4.26
pytest>=2.6.4
pytest-cache>=1.0
pytest-pep8>=1.0.6
requests>=2.4.3
six>=1.8.0
websocket-client>=0.21.0
tornado>=4.0
toro==0.7.0
